+++ lib/Braintree/Discount.php
@@ -1,4 +1,15 @@
 <?php
 class Braintree_Discount extends Braintree_Modification
 {
+    public static function all()
+    {
+        $response = Braintree_Http::get('/discounts');
+
+        $discounts = array("discount" => $response['discounts']);
+
+        return Braintree_Util::extractAttributeAsArray(
+            $discounts,
+            'discount'
+        );
+    }
 }