+++ tests/integration/TransactionTest.php
@@ -4,6 +4,78 @@ require_once realpath(dirname(__FILE__)) . '/SubscriptionTestHelper.php';
 
 class Braintree_TransactionTest extends PHPUnit_Framework_TestCase
 {
+    function testCloneTransaction()
+    {
+        $result = Braintree_Transaction::sale(array(
+            'amount' => '100.00',
+            'orderId' => '123',
+            'creditCard' => array(
+                'number' => '5105105105105100',
+                'expirationDate' => '05/2011',
+            ),
+            'customer' => array(
+                'firstName' => 'Dan',
+            ),
+            'billing' => array(
+                'firstName' => 'Carl',
+            ),
+            'shipping' => array(
+                'firstName' => 'Andrew',
+            )
+      ));
+      $this->assertTrue($result->success);
+      $transaction = $result->transaction;
+
+      $cloneResult = Braintree_Transaction::cloneTransaction($transaction->id, array('amount' => '123.45', 'options' => array('submitForSettlement' => false)));
+      Braintree_TestHelper::assertPrintable($cloneResult);
+      $this->assertTrue($cloneResult->success);
+      $cloneTransaction = $cloneResult->transaction;
+      $this->assertEquals('Dan', $cloneTransaction->customerDetails->firstName);
+      $this->assertEquals('Carl', $cloneTransaction->billingDetails->firstName);
+      $this->assertEquals('Andrew', $cloneTransaction->shippingDetails->firstName);
+      $this->assertEquals('510510******5100', $cloneTransaction->creditCardDetails->maskedNumber);
+      $this->assertEquals('authorized', $cloneTransaction->status);
+    }
+
+
+    function testCloneTransactionAndSubmitForSettlement()
+    {
+        $result = Braintree_Transaction::sale(array(
+            'amount' => '100.00',
+            'creditCard' => array(
+                'number' => '5105105105105100',
+                'expirationDate' => '05/2011',
+            )
+        ));
+
+      $this->assertTrue($result->success);
+      $transaction = $result->transaction;
+
+      $cloneResult = Braintree_Transaction::cloneTransaction($transaction->id, array('amount' => '123.45', 'options' => array('submitForSettlement' => true)));
+      $cloneTransaction = $cloneResult->transaction;
+      $this->assertEquals('submitted_for_settlement', $cloneTransaction->status);
+    }
+
+    function testCloneWithValidations()
+    {
+        $result = Braintree_Transaction::credit(array(
+            'amount' => '100.00',
+            'creditCard' => array(
+                'number' => '5105105105105100',
+                'expirationDate' => '05/2011'
+            )
+      ));
+      $this->assertTrue($result->success);
+      $transaction = $result->transaction;
+
+      $cloneResult = Braintree_Transaction::cloneTransaction($transaction->id, array('amount' => '123.45'));
+      $this->assertFalse($cloneResult->success);
+
+      $baseErrors = $cloneResult->errors->forKey('transaction')->onAttribute('base');
+
+      $this->assertEquals(Braintree_Error_Codes::TRANSACTION_CANNOT_CLONE_CREDIT, $baseErrors[0]->code);
+    }
+
     function testSale()
     {
         $result = Braintree_Transaction::sale(array(